+++ include/drawing.h
@@ -6,7 +6,7 @@
 /*   By: tkuramot <tkuramot@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/02 20:37:11 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/11/27 00:29:42 by tkuramot         ###   ########.fr       */
+/*   Updated: 2023/11/28 14:31:25 by tkuramot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,11 +21,16 @@ void	frame_buffer_apply(t_mlx_data *mlx_data);
 void	frame_buffer_destroy(t_mlx_data *mlx_data);
 void	my_mlx_pixel_put(t_frame_buffer *frame_buffer,
 				int y, int x, int color);
+unsigned int	extract_color_from_texture(t_texture *texture, int y, int x);
 void	draw_square(t_mlx_data *mlx_data,
 			t_vec2i start, t_vec2i end, int color);
 void	render_floor(t_mlx_data *mlx_data, int color);
 void	render_ceiling(t_mlx_data *mlx_data, int color);
 void	render_wall_vertical_line(t_mlx_data *mlx_data,
 			int x, int line_height, int color);
+t_texture *get_side_texture(t_world *world, t_dda *dda);
+void	calculate_line(t_wall_line *line, t_dda *dda);
+void	calculate_texture_position(t_wall_line *line, t_world *world, t_dda *dda, t_texture *texture);
+void	render_textured_wall_vertical_line(t_world *world, t_texture *texture, t_wall_line *line, int window_x);
 
 #endif
