+++ src/drawing/render_floor_and_ceiling.c
@@ -1,22 +1,27 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   dda.h                                              :+:      :+:    :+:   */
+/*   render_floor_and_ceiling.c                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tkuramot <tkuramot@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/13 00:58:24 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/11/28 14:28:23 by tkuramot         ###   ########.fr       */
+/*   Created: 2023/11/28 14:15:13 by tkuramot          #+#    #+#             */
+/*   Updated: 2023/11/28 14:15:15 by tkuramot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef DDA_H
-# define DDA_H
+#include "drawing.h"
+#include "type.h"
 
-# include "type.h"
+void	render_floor(t_mlx_data *mlx_data, int color)
+{
+	draw_square(mlx_data,
+		(t_vec2i){0, 0}, (t_vec2i){WINDOW_WIDTH, WINDOW_HEIGHT / 2}, color);
+}
 
-void	prepare_dda(t_world *world, t_dda *dda, int x);
-void	perform_dda(t_world *world, t_dda *dda);
-void	calculate_dist_camera_plane_to_wall(t_dda *dda);
-
-#endif
+void	render_ceiling(t_mlx_data *mlx_data, int color)
+{
+	draw_square(mlx_data,
+		(t_vec2i){0, WINDOW_HEIGHT / 2},
+		(t_vec2i){WINDOW_WIDTH, WINDOW_HEIGHT - 1}, color);
+}
