+++ src/game/game_loop.c
@@ -6,7 +6,7 @@
 /*   By: tkuramot <tkuramot@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/02 20:37:11 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/11/28 14:00:59 by tkuramot         ###   ########.fr       */
+/*   Updated: 2023/11/28 14:32:18 by tkuramot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,78 +54,11 @@ void	key_handler(t_world *world)
 		move_rightward(world);
 }
 
-unsigned int	extract_color_from_texture(t_texture *texture, int y, int x)
-{
-	char	*dst;
-
-	dst = texture->addr
-		+ (y * texture->line_length
-			+ x * (texture->bits_per_pixel / 8));
-	return (*(unsigned int *)dst);
-}
-
-void	render_textured_wall_vertical_line(t_world *world, t_dda *dda, t_texture *texture, double dist_camera_plane_to_wall, int window_x)
-{
-	double	wall_x;
-
-	if (dda->hit_side == WEST || dda->hit_side == EAST)
-		wall_x = world->player.precise_pos.y + dist_camera_plane_to_wall * dda->ray_dir.y;
-	else
-		wall_x = world->player.precise_pos.x + dist_camera_plane_to_wall * dda->ray_dir.x;
-	wall_x -= floor(wall_x);
-
-	int		texture_x;
-	int		texture_y;
-	int		window_y;
-	double	temp_texture_y;
-	double	step;
-	int		line_height;
-	int		line_start;
-	int		line_end;
-	unsigned int	color;
-
-	texture_x = (int)(wall_x * (double)texture->width);
-	if (dda->hit_side == WEST || dda->hit_side == NORTH)
-		texture_x = texture->width - texture_x - 1;
-	line_height = (int)(WINDOW_HEIGHT / dist_camera_plane_to_wall);
-	line_start = -line_height / 2 + WINDOW_HEIGHT / 2;
-	if (line_start < 0)
-		line_start = 0;
-	line_end = line_height / 2 + WINDOW_HEIGHT / 2;
-	if (line_end >= WINDOW_HEIGHT)
-		line_end = WINDOW_HEIGHT - 1;
-	step = 1.0 * texture->height / line_height;
-	window_y = line_start;
-	temp_texture_y = (int)(line_start - WINDOW_HEIGHT / 2 + line_height / 2) * step;
-	while (window_y < line_end)
-	{
-		texture_y = (int)fmod(temp_texture_y, texture->height - 1);
-		temp_texture_y += step;
-		color = extract_color_from_texture(texture, texture_y, texture_x);
-		my_mlx_pixel_put(&world->mlx_data.frame_buffer, window_y, window_x, color);
-		window_y++;
-	}
-}
-
-t_texture *get_side_texture(t_world *world, t_dda *dda)
-{
-	if (dda->hit_side == NORTH)
-		return (&world->mlx_data.textures[NORTH]);
-	else if (dda->hit_side == SOUTH)
-		return (&world->mlx_data.textures[SOUTH]);
-	else if (dda->hit_side == WEST)
-		return (&world->mlx_data.textures[WEST]);
-	else if (dda->hit_side == EAST)
-		return (&world->mlx_data.textures[EAST]);
-	else
-		return (NULL);
-}
-
 int	game_loop(t_world *world)
 {
-	t_dda	dda;
-	int		window_x;
-	double	dist_camera_plane_to_wall;
+	t_dda		dda;
+	int			window_x;
+	t_wall_line	line;
 
 	key_handler(world);
 	render_floor(&world->mlx_data, world->floor_color);
@@ -135,8 +68,10 @@ int	game_loop(t_world *world)
 	{
 		prepare_dda(world, &dda, window_x);
 		perform_dda(world, &dda);
-		dist_camera_plane_to_wall = get_dist_camera_plane_to_wall(&dda);
-		render_textured_wall_vertical_line(world, &dda, get_side_texture(world, &dda), dist_camera_plane_to_wall, window_x);
+		calculate_dist_camera_plane_to_wall(&dda);
+		calculate_line(&line, &dda);
+		calculate_texture_position(&line, world, &dda, get_side_texture(world, &dda));
+		render_textured_wall_vertical_line(world, get_side_texture(world, &dda), &line, window_x);
 		window_x++;
 	}
 	frame_buffer_apply(&world->mlx_data);
